Move startEndpoint from ServerEndpoint to ServerEndpointRecipe
