+++ framework/src/play-test/src/main/scala/play/api/test/ServerEndpoint.scala
@@ -11,13 +11,10 @@ import com.typesafe.sslconfig.ssl.{ FakeKeyStore, FakeSSLTools }
 
 import akka.annotation.ApiMayChange
 
-import play.api.{ Application, Mode }
 import play.api.test.ServerEndpoint.ClientSsl
 import play.core.ApplicationProvider
 import play.core.server.ServerConfig
 import play.server.api.SSLEngineProvider
-
-import scala.util.control.NonFatal
 
 /**
  * Contains information about which port and protocol can be used to connect to the server.
@@ -45,54 +42,6 @@ import scala.util.control.NonFatal
 @ApiMayChange object ServerEndpoint {
   /** Contains SSL information for a client that wants to connect to a [[ServerEndpoint]]. */
   final case class ClientSsl(sslContext: SSLContext, trustManager: X509TrustManager)
-
-  /**
-   * Starts a server by following a [[ServerEndpointRecipe]] and using the
-   * application provided by an [[ApplicationFactory]]. The server's endpoint
-   * is passed to the given `block` of code.
-   */
-  def startEndpoint[A](endpointRecipe: ServerEndpointRecipe, appFactory: ApplicationFactory): (ServerEndpoint, AutoCloseable) = {
-    val application: Application = appFactory.create()
-
-    // Create a ServerConfig with dynamic ports and using a self-signed certificate
-    val serverConfig = {
-      val sc: ServerConfig = ServerConfig(
-        port = endpointRecipe.configuredHttpPort,
-        sslPort = endpointRecipe.configuredHttpsPort,
-        mode = Mode.Test,
-        rootDir = application.path
-      )
-      val patch = endpointRecipe.serverConfiguration
-      sc.copy(configuration = sc.configuration ++ patch)
-    }
-
-    // Initialize and start the TestServer
-    val testServer: play.api.test.TestServer = new play.api.test.TestServer(
-      serverConfig, application, Some(endpointRecipe.serverProvider)
-    )
-
-    val runSynchronized = application.globalApplicationEnabled
-    if (runSynchronized) {
-      PlayRunners.mutex.lock()
-    }
-    val stopEndpoint = new AutoCloseable {
-      override def close(): Unit = {
-        testServer.stop()
-        if (runSynchronized) {
-          PlayRunners.mutex.unlock()
-        }
-      }
-    }
-    try {
-      testServer.start()
-      val endpoint: ServerEndpoint = endpointRecipe.createEndpointFromServer(testServer)
-      (endpoint, stopEndpoint)
-    } catch {
-      case NonFatal(e) =>
-        stopEndpoint.close()
-        throw e
-    }
-  }
 
   /**
    * An SSLEngineProvider which simply references the values in the
